backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.25
bcrypt==4.1.2
redis==5.0.1
celery==5.3.4
